Added `de/registrations` module
