+++ crates/bevy_reflect/src/serde/de/mod.rs
@@ -1,3 +1,6 @@
+pub use registrations::*;
+mod registrations;
+
 use crate::serde::SerializationData;
 use crate::{
     ArrayInfo, DynamicArray, DynamicEnum, DynamicList, DynamicMap, DynamicSet, DynamicStruct,
@@ -238,54 +241,6 @@ impl<'de> Deserialize<'de> for Ident {
         }
 
         deserializer.deserialize_identifier(IdentVisitor)
-    }
-}
-
-/// A deserializer for type registrations.
-///
-/// This will return a [`&TypeRegistration`] corresponding to the given type.
-/// This deserializer expects a string containing the _full_ [type path] of the
-/// type to find the `TypeRegistration` of.
-///
-/// [`&TypeRegistration`]: TypeRegistration
-/// [type path]: crate::TypePath::type_path
-pub struct TypeRegistrationDeserializer<'a> {
-    registry: &'a TypeRegistry,
-}
-
-impl<'a> TypeRegistrationDeserializer<'a> {
-    pub fn new(registry: &'a TypeRegistry) -> Self {
-        Self { registry }
-    }
-}
-
-impl<'a, 'de> DeserializeSeed<'de> for TypeRegistrationDeserializer<'a> {
-    type Value = &'a TypeRegistration;
-
-    fn deserialize<D>(self, deserializer: D) -> Result<Self::Value, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        struct TypeRegistrationVisitor<'a>(&'a TypeRegistry);
-
-        impl<'de, 'a> Visitor<'de> for TypeRegistrationVisitor<'a> {
-            type Value = &'a TypeRegistration;
-
-            fn expecting(&self, formatter: &mut Formatter) -> fmt::Result {
-                formatter.write_str("string containing `type` entry for the reflected value")
-            }
-
-            fn visit_str<E>(self, type_path: &str) -> Result<Self::Value, E>
-            where
-                E: Error,
-            {
-                self.0.get_with_type_path(type_path).ok_or_else(|| {
-                    Error::custom(format_args!("No registration found for `{type_path}`"))
-                })
-            }
-        }
-
-        deserializer.deserialize_str(TypeRegistrationVisitor(self.registry))
     }
 }
